More comment fixes - consume whitespace before block comment
Previously, when block comments were preserved in token stream, and
created Comment nodes in the grammar, it was fine to have extra
TERMINATORs and INDENTs in the stream as well - often preceding a
HERECOMMENT. Since it was part of the grammar, HERECOMMENTs e.g. in
object definitions were fine, because it was INDENT and then a
meaningful HERECOMMENT token. But now that we remove all HERECOMMENTs in
rewriter, this would leave an empty INDENTs followed by TERMINATORs,
which do not pass the grammar.